\name{Regression.Sweep.PCR.Config-class}
\Rdversion{1.1}
\docType{class}
\alias{Regression.Sweep.PCR.Config-class}

\title{Class \code{"Regression.Sweep.PCR.Config"}}
\description{
Configuration class for PCR sweep operation
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("Regression.Sweep.PCR.Config", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{n}:}{Object of class \code{"OptionalNumeric"}, indicating, in this derived class, the maximum number of PC's to perform the PCR sweep for.}
  }
}
\section{Extends}{
Class \code{"Regression.Sweep.Config"}, directly.
}
\section{Methods}{
  \describe{
    \item{Regression.Sweep.Fit}{\code{signature(object = "Regression.Sweep.PCR.Config")}: ... }
   }
}

\author{
Mansour T.A. Sharabiani, Alireza S. Mahani
}

\keyword{classes}
